Move SwipeableWord to platform code
